DT-6387 mqtt topic fecthing to utils, less variables passed
